Create README.md
This program will prompt for a URL, read the XML data from that URL using urllib and then parse and extract the comment counts from the XML data, compute the sum of the numbers in the file and enter the sum of numbers in the file. Here  we are to look through all the <comment> tags and find the <count> values sum the numbers.